feat: created forgotPassword token and reset password routes and now able to update the forgotten password

+++ controllers/bootcamp.js
@@ -34,10 +34,20 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route POST /api/bootcamp
 // @access private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+    //add user to req.body
+    req.body.user = req.user.id
+
+    //check for published bootcamp
+    const publishedBootcamp = await Bootcamp.findOne({user: req.user.id})
+
+    //if the user is not admin they can only add one bootcamp
+    if(publishedBootcamp && req.user.role !== 'admin'){
+        return next( new ErrorResponse(`User with the ID ${req.user.id} has already published a bootcamp`, 400));
+    }
 
     const bootcamp = await Bootcamp.create(req.body);
-    res.status(201).json({ success: true, data: bootcamp });
 
+    res.status(201).json({ success: true, data: bootcamp });
 });
 
 //@desc update bootcamp
@@ -45,14 +55,20 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @access private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    })
+    let bootcamp = await Bootcamp.findById(req.params.id)
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp has not been found with the id of ${req.params.id}`, 404));
     }
+    //make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User with ID ${req.params.id} is not authorized to update the bootcamp`, 401))
+    }
+
+   bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true
+   })
 
     res.status(200).json({ success: true, data: bootcamp })
 
@@ -63,14 +79,18 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 // @access private
 exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 
-    const bootcamp = await Bootcamp.findById(req.params.id);
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp has not been found with the id of ${req.params.id}`, 404));
     }
 
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User with ID ${req.params.id} is not authorized to delete the bootcamp`, 401))
+    }
+
     await bootcamp.deleteOne();
-    res.status(200).json({ success: true, msg: ` Deleted a bootcamp ${req.params.id}` });
+    res.status(200).json({ success: true, msg: ` Deleted a bootcamp ${req.params.id}`, data: {} });
 
 });
 
@@ -110,9 +130,16 @@ exports.bootcampPhotoUpload = asyncHandler( async (req, res, next) => {
         return next( new ErrorResponse( `Bootcamp not found with the id of ${req.params.id}`), 404)
     }
 
+    //make sure user is bootcamp owner
+     if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User with ID ${req.params.id} is not authorized to delete the bootcamp`, 401))
+    }
+
     if(!req.files){
        return next( new ErrorResponse('Please upload a file',400))
     }
+
+
 
     const file = req.files.file;
